Move installWebpackConfig to blitz-rpc
